refactor: 코드 리뷰 반영
- ColumnNameProperty로 추상화
- TableName클래스 패키지 이동
